# contracts/openapi/operations.yaml
openapi: 3.1.0
info:
  title: Qurban Operations API
  version: 1.0.0-draft
  description: |
    Phase 1 private Operations contract. It specifies OIDC-backed sessions,
    permission-gated commands, and minimized administrative views. Event and
    Offering routes are implemented; later Operations groups remain draft.
servers:
  - url: https://api.example.invalid
    description: Deployment-specific base URL
security: []
...
paths:
...
  /api/operations/v1/payments:
    get:
      tags: [Payments]
      operationId: listPayments
      summary: List payments
      description: Cursor pagination defaults to submitted_at descending, then id ascending.
      security: [{ OperationsSession: [] }]
      x-required-permission: payment.read
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          required: false
          schema:
            { type: string, enum: [SUBMITTED, VERIFIED, REJECTED, REFUNDED] }
      responses:
        "200": { $ref: "#/components/responses/PaymentListResponse" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/operations/v1/payments/{payment_id}:
    get:
      tags: [Payments]
      operationId: getPayment
      summary: Get an Operations payment view
      security: [{ OperationsSession: [] }]
      x-required-permission: payment.read
      parameters:
        - $ref: "#/components/parameters/PaymentID"
        - $ref: "#/components/parameters/RequestID"
      responses:
        "200": { $ref: "#/components/responses/PaymentResponse" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    OperationsSession:
      type: apiKey
      in: cookie
      name: __Host-operations_session
      description: Secure, HttpOnly, SameSite=Lax server-side session. The raw token is never a response body field.
  headers:
    RequestID:
      description: Accepted X-Request-ID or a generated UUID request correlation identifier.
      required: true
      schema:
        {
          type: string,
          maxLength: 64,
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
        }
...
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      schema:
        {
          type: string,
          minLength: 1,
          maxLength: 64,
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
        }
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, minLength: 1, maxLength: 512 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
    PaymentID:
      name: payment_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    PaymentResponse:
      description: Operations payment view.
      headers: { X-Request-ID: { $ref: "#/components/headers/RequestID" } }
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  required: [data],
                  properties:
                    { data: { $ref: "#/components/schemas/Payment" } },
                },
            },
        }
    PaymentListResponse:
      description: Payment page.
      headers: { X-Request-ID: { $ref: "#/components/headers/RequestID" } }
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  required: [data, page],
                  properties:
                    {
                      data:
                        {
                          type: array,
                          items: { $ref: "#/components/schemas/Payment" },
                        },
                      page: { $ref: "#/components/schemas/PageInfo" },
                    },
                },
            },
        }
...
    InvalidRequest:
      description: Malformed request or correlation header.
      headers: { X-Request-ID: { $ref: "#/components/headers/RequestID" } }
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
    Unauthenticated:
      description: Missing, invalid, expired, or revoked credential.
      headers: { X-Request-ID: { $ref: "#/components/headers/RequestID" } }
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
    Forbidden:
      description: CSRF, Origin, or permission policy rejected the request.
      headers: { X-Request-ID: { $ref: "#/components/headers/RequestID" } }
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
    NotFound:
      description: Resource not found.
      headers: { X-Request-ID: { $ref: "#/components/headers/RequestID" } }
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
...
    InternalError:
      description: Unexpected error; no SQL, stack trace, credentials, or personal data is disclosed.
      headers: { X-Request-ID: { $ref: "#/components/headers/RequestID" } }
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
...
  schemas:
...
    PartySummary:
      type: object
      additionalProperties: false
      required: [id, display_name]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
...
    Payment:
      type: object
      additionalProperties: false
      required:
        [
          id,
          event_id,
          payment_ref,
          payer,
          amount_minor,
          currency_code,
          method,
          status,
          submitted_at,
        ]
      properties:
        id: { type: string, format: uuid }
        event_id: { type: string, format: uuid }
        payment_ref: { type: string }
        purchase_id: { type: string, format: uuid }
        payer: { $ref: "#/components/schemas/PartySummary" }
        amount_minor: { type: integer, minimum: 1 }
        currency_code: { type: string, pattern: "^[A-Z]{3}$" }
        method: { type: string }
        status:
          { type: string, enum: [SUBMITTED, VERIFIED, REJECTED, REFUNDED] }
        evidence:
          type: object
          additionalProperties: false
          properties:
            filename: { type: string }
            media_type:
              { type: string, enum: [image/jpeg, image/png, application/pdf] }
            size_bytes: { type: integer, minimum: 1, maximum: 10485760 }
        submitted_at: { type: string, format: date-time }
        verified_at: { type: string, format: date-time }
        rejection_reason: { type: string }
...
    PageInfo:
      type: object
      additionalProperties: false
      required: [limit]
      properties:
        next_cursor: { type: string, maxLength: 512 }
        limit: { type: integer, minimum: 1, maximum: 100 }
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id, details]
          properties:
            code:
              type: string
              enum:
                [
                  invalid_request,
                  validation_failed,
                  unauthenticated,
                  forbidden,
                  not_found,
                  method_not_allowed,
                  stale_version,
                  state_conflict,
                  quota_unavailable,
                  idempotency_conflict,
                  evidence_too_large,
                  request_too_large,
                  unsupported_media_type,
                  rate_limited,
                  internal_error,
                  service_unavailable,
                ]
            message: { type: string }
            request_id: { type: string, maxLength: 64 }
            details: { type: object, additionalProperties: true }